+++ src/java/com/threerings/msoy/web/client/SwiftlyService.java
@@ -32,7 +32,7 @@ public interface SwiftlyService extends RemoteService
     /**
      * Creates a new SwiftlyProject for the member in the supplied WebCreds
      */
-    public SwiftlyProject createProject (WebCreds creds, String projectName, int projectType,
+    public SwiftlyProject createProject (WebCreds creds, String projectName, byte projectType,
                                          boolean remixable)
         throws ServiceException;
